feat(userspace): musl ABI validation for stat/uname syscalls
hello_musl.c updates:
- Add struct stat validation (MUSL-STAT-OK marker)
- Add struct utsname validation (MUSL-UNAME-OK marker)
- Print layout details for ABI verification
- Total markers: 6 (was 4)

Shell and syscall integration:
- Update shell for capability-aware file operations
- Syscall wrapper adjustments for cap_id propagation

Rebuilt musl_test.elf and hello_musl.elf with new validation.

Part of D2-ABI-STAT-LAYOUT and D2-TST design queue closure.

+++ userspace/src/shell.rs
@@ -585,9 +585,9 @@ fn do_stat(args: &[u8]) {
     println("");
 
     print("  Size: ");
-    print_int(st.size as i64);
+    print_int(st.size);
     print("  Blocks: ");
-    print_int(st.blocks as i64);
+    print_int(st.blocks);
     println("");
 
     print("Device: ");
